Add alpha fading for windows
Remove old "win alpha" command. Add the following commands:

  win alpha0 val time
  win alpha1 val time
  win alpham mode
  win fade

The first two set the new alpha0 or alpha1 value. If the time
is zero, the value is immediately set, otherwise the change is
registered to be triggered later with "win fade".

The third command sets the new alpha mode (0..2).

The fourth command actually performs the alpha fading. All
windows on all displays that have an alpha change registered
with "win alpha0" or "win alpha1" now gradually change their
alpha value to the given target value in the given time. This
function can be interrupted with Ctl-C and resumed with another
call to "win fade".