Rename armStoneA8 configuration to fss5pv210
The code for armStoneA8 ist already prepared to also work on
other F&S boards based on the S5PV210 MCU like the NetDCU14.
To reflect this multi-board purpose, call the platform fss5pv210
instead of armstonea8 now. So you have to call

  make fss5pv210

or

  make fss5pv210_config

to configure for any of these boards, including armStoneA8.

- Move include/configs/armstonea8.h to include/configs/fss5pv210.h;
  rename some defines
- Move board/F+S/armstonea8 directory to board/F+S/fss5pv210
- Rename board/F+S/armstonea8.c to board/F+S/fss5pv210.c; modify
  board/F+S/fss5pv210/Makefile to reflect that change
- Modify some occurances of armstonea8 in board/F+S/fss5pv210
- Modify boards.cfg to reflect the name change