Add functions get_board_name() and get_sys_prompt()
get_board_name() returns a pointer to the name of the board.
The default implementation returns CONFIG_SYS_BOARDNAME, or
"u-boot" if CONFIG_SYS_BOARDNAME is not defined.

get_sys_prompt() returns a pointer to the system prompt string.
The default implementation returns CONFIG_SYS_PROMPT, or
" # " appended to the board name CONFIG_SYS_BOARDNAME if
CONFIG_SYS_PROMPT is not defined.