Add commands update and install to trigger installation
These two commands trigger the automatic update/installation
process that can also be automatically performed at system
start.

update [updscript [updcheck [updaddr]]]
install [instscript [instcheck [instaddr]]]

You can give the script name, devices to check and the load
address on the command line. All arguments that are not given
on the command line are taken from the environment variables
with the same name.

While doing this change, also the comments that are output
while trying to find the script on the different devices were
improved.

This commit resolves bugtracker issue #1254.