Allow board specific serial port, set sercon variable
Different boards may have a different default serial port
(debug port), e.g. armStoneA8 and NetDCU14. Provide an own
function default_serial_console() that sets the correct port
for this board (as reported by NBoot via the nboot_args).

In addition set an environment variable sercon to this port
in board_late_init(), so that it can be used in the different
bootargs settings. We can't set this value in the builtin
bootargs, so we set the bootargs to empty. And if bootargs is
not set, run bootubi variable by default in board_late_init().

This commit resolves bugtracker issue #1260.