Set bootcmd to load to loadaddr instead of fix 0x41000000
The kernel should be loaded to loadaddr, not a fix address.
loadaddr is dynamically set depending on the RAM size of the
board. On boards that use both DRAM banks, we must load to
the lower bank or Linux won't see the whole RAM. 0x41000000
was in the upper bank, so Linux only saw half of the RAM.

This commit resolves bugtracker issue #1164.