Patch serial port in zImage decompressor
Unfortunately the Linux configuration sets a fix port
to use while decompressing the zImage. To allow this code
to use the right serial port of the underlying board, we
patch this port in the zImage code before executing the
zImage. This requires the port to be set as ttySACn in the
first part of the image, where we modify n. The value of
CONFIG_SYS_PATCH_TTY also determines the number of bytes
of the image that is checked for the patch. So if you set
this value to 0x4000, the first 16KB of the image are checked.