Allow different ECC modes by virtually splitting NAND
On Samsung MCUs, the stepstone loader (NBoot) has to use
8-Bit-ECC, while U-Boot, Kernel and Root Filesystem are using
a standard 1-Bit-ECC. So to be able to access NBoot from
U-Boot (or Linux), we have to support both ECC modes. The
implementation from PicoMOD6/PicoCOM3 (U-Boot 1.3.4, Kernel
2.6.28) does not work with newer versions anymore. So we have
to set up a new clean way to implement this.

We do this by virtually splitting the NAND into two devices.
One with NBoot and one with the remaining data. On the first,
the size of the device is restricted to the NBoot size, on
the second the beginning of the device of the size of NBoot
is ignored (skipped). For each device we can define an own
ECC strategy.

At the same time we remove the CONFIG_NAND_NBOOT setting that
should have done the same but was not as flexible.

Having two devices allows to have a software protection for
each of them. So a new command "nand protect" replaces the
previous command "nand nboot protect".

The ECC layout can now be moved from the board-specific code
to the more general Samsung NAND code.

This commit resolves bugtracker issues #1256 and #1258.