Allow dynamic mtdparts sizes depending on NAND flash type
Up to now, the mtdparts defaults setting was a fix setting
given in the U-Boot configuration. However this setting depends
on the block size of the NAND chip. On boards that allow NAND
chips with different block sizes (e.g. PicoMOD6, PicoCOM3),
this should be set depending on the actually mounted NAND chip.

If the default setting for mtdparts is empty, the mtdparts
command now calls function board_get_mtdparts_default() where
we can dynamically return a setting reflecting the current NAND
setup. In addition we also set the mtdparts environment variable
if it is empty. This saves a call to mtdparts default if the
board is initialized for the first time.

This commit resolves bugtracker issue #1257.