Reduce environment size on large NAND blocksizes
We always had the environment size set to the size of one NAND
block. On NAND chips with 128KB blocks, we waste a lot of
malloc space because the environment is located in the heap
and may need the same size a second time when saving. 128KB
is way too much for the environment anyway. So use 16KB like
on small NAND blocks.