Collect all possible command configurations in one place
Update fss5pv210.h and fss3c64xx.h to show all possible command
configurations in one place in the file. We need support for
FAT and EXT2 filesystems because SD cards and USB sticks may
be formatted this way. But we don't need support for UBIFS
and JFFS2 filesystems. We just need to be able to store those
images as a whole, but we don't need access to single files
within these images.

There is a new bootz command to load zImages. Convert environment
variables to use it instead of bootm.

Enable flat device tree support.

The following commands are now activated by default:
- CONFIG_CMD_UPDATE
- CONFIG_CMD_BOOTZ
- CONFIG_CMD_UNZIP
- CONFIG_OF_FDT

The following commands are now deactivated by default:
- CONFIG_CMD_ELF
- CONFIG_CMD_ITEST
- CONFIG_CMD_UBIFS
- CONFIG_CMD_JFFS2