Improve device description output (interface, indendation)
Device descriptions assumed to be appended to a device name that
is already output from the caller. However the indendation was
not consistent accross the different modules. So redefine to
output info in separate lines with two spaces indendation. This is
how it is done in other modules.

Also move the interface description from fs/fat/fat.c to the general
device description in disk/part.c.