Add wildcard module to support wildcards in file names
Up to now, each filesystem implemented its own functions to walk
the directory tree and find file names. This resulted in a lot of
duplicated code and all these implementations were rather simple
and therefore could not handle any wildcard characters. In fact
every filesystem implemented its own version of the ls command.

The new wildcard module is designed to do this in a more generic
way. It contains functions to list directories in 'ls' style
and read and write single files. The actual data accesses are
done by calling a set of filesystem specific functions, but the
path resolution and pattern matching is done in the wildcard
module.

The module supports the following wildcards:

'?' matches a single arbitrary character
'*' matches zero or more arbitrary characters

The ls command will show several file lists if the path matches
more than one directory.

Wildcards are also allowed if reading or writing a file, but then
the match must be unambiguous. So here the wildcard is only useful
to save typing on long file names.

If a filesystem wants to support the wildcard module, it has to
provide the following functions:

alloc_dir(): Allocate a structure to represent a directory
free_dir():  Free the above structure
get_fileinfo(): Return info about the next file in the directory
read_file(): Read a file
write_file(): Write a file
get_symlink(): Return the symbolic link (relative path)

Remark: symbolic links are not fully supported yet