New FAT/VFAT implementation
The previous FAT implementation used three large static buffers
of 64KB and another buffer on the heap. Accessing files with long
names did never work satisfyingly. Reading and writing of files
and directories was partly inefficient.

The new implementation only uses two static buffers with 4KB and 3KB
respectively (configurable). Access to long filenames is drastically
improved, all filename handling and the ls command is done by the
wildcard module, thus allowing the use of wildcards in paths and
filenames.

The file fs/fat/file.c is not required anymore and can be removed in
the future.

Remark: Write support is not yet implemented.