Optimize speed for SD cards and fix high speed issues
The S3C6410 and S5PV210 platforms do support the HISPD bit in
the host controller (despite the fact that it is called OUTEDGEINV
in the documentation), but timing changes only for reading, not
writing in high speed. Therefore it does not make sense to set
this bit in high speed mode, and it also does not work. So
re-activate the SDHCI_QUIRK_NO_HISPD_BIT.

- Use actual base_clock (and base_clock/128) in s3c_sdhci_init()
  and s5p_sdhci_init() respectively instead of fix 52MHz and 400kHz.
- In mmc.c, set the host clock immediately after changing the
  SD-card speed; don't issue other commands (set bus width) in
  between.
- Correct description for feedback clock and base clock in
  s5p_sdhci.c.

This commit resolves bugtracker issue #1526.