Patch serial port for zImage decompressor on Vybrid
When booting a zImage with bootm or bootz, the image itself contains
a decompressor for the kernel. This decompressor needs a serial port
to report any problems. However the serial debug port differs on
different F&S boards. Therefore we have modified the Linux decompressor
so that the beginning of the image now contains the string "ttymxcN"
(or CONFIG_SYS_SERCON_NAME in general). If this string is found, the
number N at the end is patched to the port number found in $sercon.
So the decompressor will report its messages on the same port as
set in $sercon.

Also fix bootubi and bootubidhcp in the default environment to use
$sercon instead of ttymxc1.