If set, read MAC addresses from OTP fuses on Vybrid
For each FEC device that is registered and no ethaddr environment
variable is set, read a MAC adddress from the OTP fuses and set the
appropriate ethaddr environment variable. If not enough MAC addresses
are programmed in OTP, fill the remaining addresses with default
addesses starting at CONFIG_ETHADDR_BASE. Especially if no MAC address
is programmed in the fuses, the addresses are set from the default
address as before.

- Add set_fs_ethaddr() to set the next MAC address either from MAC0
  in OTP, MAC1 in OTP or from the default address CONFIG_ETHADDR_BASE
- Add get_otp_mac() to read a MAC address from OTP
- Add eth_setenv_enetaddr_by_index() to set the appropriate ethaddr
  environment variable given by an index