Remove setting of fec_mac from fsvybrid bootargs
The improved FEC handling on Vybrid allows Linux to take the MAC
address directly from the FEC controller. It is not necessary anymore
to pass the MAC address on the command line. In fact if it is passed
on the command line, it overrides the one already in the controller,
and this is not what we want.