Backport nand command handling from u-boot-2013.07
If the size of nand read or nand write commands is not given, the size
is taken from the partition size. We have to take care that reading
or writing does not exceed into the next partition if there are bad
blocks. This is handled in newer U-Boot versions already, so backport
the behaviour from there.

We also take some improvements in lock handling from there.

- Handle maxsize when parsing arguments of nand commands. Adjust the
  size about the amount of bad blocks when size is not given and
  taken from the partition/device size.
- Handle limit and actual value in nand_write_skip_bad() and
  nand_read_skip_bad(); this also requires a modification in readenv()
  in env_nand.c where nand_read_skip_bad() is called, too.
- Allow .allexcept extension to nand unlock command.
- Check lock-tight status before locking or unlocking.
- Add optional nand torture command.
- Use new setenv_hex() in nand_print_and_set_info(), replace
  setenv_addr() with setenv_hex() in cmd_nvedit.c, implement
  setenv_addr() as a macro using setenv_hex() in common.h.

Please note that we do not take the newer NAND access stuff introduced
in u-boot-2013.07. Here we only take the code up to u-boot-2013.04.