Fix NAND ECC correction on fss5pv210
Function s5p_nand_correct_data() used the ECC that was computed
just ago as comparison for the current ECC. Obviously this was always
correct and never detected any errors. But instead it must use the
ECC that was read from the OOB area to make sense and actually detect
any bitflips.