Do not reduce NAND size by skip size on fsvybrid and fs3c64xx
To allow handling the NBoot region with a different ECC strategy, we
can split the NAND MTD device in two virtual devices. One for the
first part of the device (i.e. NBoot) and one for the remaining part.
The first device simply has the size of NBoot, and the second device
has a region of NBoot size that is skipped at the beginning of the
device.

If the NAND size of this second device is reduced about the skip
size, the last MTD partition (usually TargetFS) will get too small.
This is not wanted. So remove the code that did this. Now the size
of the first device will add to the overall size as we compute the
NBoot region twice. But as this is only 256K at max and the NAND
size is shown in MB, it will not change the reported value.