Fix NAND ECC correction on fss3c64xx
Function s3c_nand_correct_data() used the ECC that was computed
just ago as comparison for the current ECC. Obviously this was always
correct and never detected any errors. But instead it must use the
ECC that was read from the OOB area to make sense and actually detect
any bitflips.