Handle empty pages with bitflips in U-Boot on fsvybrid
If an empty page contains bitflips, the empty-check always failed.
This sometimes caused that UBI failed to attach to the UBI partition.

Now the empty-check accepts a small number of bitflips. If this
happens, the read function returns -EUCLEAN to make the filesystem
aware of this fact. Then UBI may schedule this block for scrubbing.

In U-Boot this will trigger a block refresh if the block/page is not
under control of a filesystem.