Improve Ethernet on fsvybrid
The new F&S Vybrid board revisions have a direct connection between
CKO1 (PTB10) and RMIICLK (PTA6). Here the RMII clock is output on
PTB10 and PTA6 is an input. This avoids a hardware bug that makes
ethernet instable when RMIICLK is directly used as the clock output.

On AGATEWAY, where CKO1 is already used differently, there is an
external oscillator for Ethernet starting with board Rev. 1.10.

When using NBoot VN11 or newer, this is detected and flag
FEAT2_RMIICLK_CKO1 is set in chFeatures2 of the NBoot args.

- Change FEAT_* define name to FEAT1_* as they are for chFeatures1.
- Add FEAT2_* defines for chFeatures2.
- Do CKO1 and RMIICLK configuration depending on FEAT2_RMIICLK_CKO1
  flag in chFeatures2.
- Move activation of RMIICLK (input or output) from board_eth_init()
  to board_init() to have it activated in any case, even if no
  ethernet is used in U-Boot. Then we do not need any modifications
  in the Linux kernel.
- Improve the comment that describes the Vybrid hardware bug.