Convert fat_read() to fat_read_at()
The fatload command has a new feature that allows reading a file from
a certain position and not only from the beginning. Therefore convert
function fat_read() to fat_read_at() and allow skipping an arbitrary
count of bytes at the beginning of the file.

The idea of the load algorithm remains the same: combine as many
subsequent sectors of the file to a large chunk. But then skip the
required number of bytes and only load the remaining part. We also
have to take care now that after reading the first few bytes, the
buffer may not be aligned correctly anymore for all subsequent full
sector reads. In this case we do single-sector reads through a one
sector auxiliary buffer.

Modifying the FAT interface also requires changing the wildcard
interface. Here wildcard_read() is also changed to wildcard_read_at().