Modify env_nand.c to work with variable environment size
The code for function saveenv() for the two cases with and without
CONFIG_ENV_OFFSET_REDUND set was merged. This new code did not work
when the environmenst size is not constant. Modify code to work
again when environment size is variable as it is on some F&S boards.