Revert moving interrupt_init() to board_init_f()
interrupt_init() uses variables, for example on Vybrid. Therefore it
can not be called in board_init_f() or it will overwrite the relocation
information that is stored after the U-Boot code in the same place as
bss and data sections. Then U-Boot is not correctly relocated to the
end of RAM. So revert the change that was introduced in u-boot-2013.10
that moved this call from board_init_r() to board_init_f(). Later
U-Boot versions have this change reverted, too.