Fix block refresh and support block refresh on fsimx6
mtd_read_oob() has to check the return value for bitflips, too, and
return -EUCLEAN or 0 in this case. Otherwise the bitflip count will
be taken as an error and for example block refresh will fail.

Support block refresh in GPMI driver on fsimx6. This needs writing
and reading the number of the currently refreshed block in the last
page of a block in the four bytes around the bad block marker.

Activate block refresh by default on fsimx6.