Hide boot strategy variables with prefix '.' instead of '_'
Current U-Boot automatically hides all environment variables that
start with a '.' character. From now on we will use this feature
to hide our boot strategy variables that would dominate the printenv
output otherwise. If you want to see the variables, just say

  printenv -a

So just replace the leading underscore of the previous version with
a leading period now. So for example to switch the kernel load method
to tftp, simply enter

  run .kernel_tftp