Add Ethernet support for PicoMODA9
PicoMODA9 uses 100 MBit/s Ethernet on (different) RMII pins and the
DP83484 PHY, not 1 GBit/s Ethernet on RGMII pins and Atheros PHY like
all the other F&S boards based on i.MX6. Add pin configuration settings
for this in fsimx6.c.

Unfortunately the Freescale driver assumed that the link type (RMII or
RGMII) can be set by a CONFIG value. So extend the driver to allow
configuration on the init call. We modify fecmxc_initialize_multi()
for this and add a version fecmxc_initialize_multi_type() that adds
the PHY type. This requires moving enum xceiver_type to netdev.h so
that it is available in the board support files.

The address of the PHY can also not be set by a CONFIG value anymore.
Set it depending on the fsimx6 board type.