Add earlyusbinit to do an early USB device scan
When U-Boot starts the USB host ports, it immediately scans for
devices. However some devices are slow and take some time before
they signal their presence, especially if their power is off until
the root hub and all intermediate hubs are activated. So at the
time when they signal their presence, the device scan is already
over. So these slow devices are not found at the first run. On the
command line, you can call "usb start" for a second time, then the
devices should be found. This is not nice, but it works. But for
actions like automatic updates or installation from an USB drive,
it is fatal if some devices are not found at the first time.

So we are doing an ugly hack here: if environment variable
earlyusbinit is set, we are doing a (silent) scan right before
the main_loop() is called in board.c, i.e. befor the boot delay.
So later, when the real scan happens, those devices will be found
immediately. Please note that this means that USB is active right
from the start! It will also slow down the start noticable.

To have a (rather) silent device scan, we have to modify usb_init()
to take an argument whether to be verbose or not. This also allows
update_mmc() in cmd_source.c to run more silently.