Add earlyusbinit to all F&S boards
Add the possibility to use earlyusbinit to all F&S architectures
(fsimx6, fsvybrid, fss5pv210, fsam335x). The default for most boards
is to not use it, as it slows down the boot process. But efusA9 needs
it activated because the USB hub on the Starterkit baseboard enables
the power to the devices only after the USB communication is running
and the HUB enables the port power. As a consequence, slow USB sticks
would be missed when checking for update or install.

The user can override this behavior by simply removing the earlyusbinit
variable.

  setenv earlyusbinit
  saveenv