Improve update command, allow arbitrary filesystem type
Update now uses the generic filesystem functions to init updates on
MMC and USB. This allows loading the script from any filesystem type
that U-Boot knows of, currently FAT and EXT2/4.

When the script is loaded successfully, update sets environment
variable "updatedev" to the device where it has found the file
(including device and partition numbers on USB and MMC). This
allows the script to continue loading from this device.

Example:
  updatecheck=mmc,usb

Update searches on mmc 0 and usb 0 for the script. Let's assume the
file is found on usb 0. Then update will set variable

  updatedev=usb 0

The script then can use this value for example with the load command:

  load $updatedev $loadaddr <filename>