Actually use new volume name format for ubifsmount
Command ubifsmount now needs "ubi", an optional number and ':'
prepended to the volume name. Now the name looks exactly the same
as on the kernel command line.

Example for volume name rootfs in the first (and only) UBI:

  ubifsmount ubi0:rootfs

The 0 can be omitted in this case.

Fix all occurences in fsvybrid.h, fsimx6.h, fss5pv210.h and fsam335x.h
as well as in the updatecheck and recovercheck default settings for
CUBEA5, AGATEWAY and HGATEWAY in fsvybrid.c.