Add rootwait for asynchronuous root devices like SD Card
Some root devices are detected in a separate thread, for example
SD cards. So if the system does not wait until this detection is
complete, it might try to mount the root filesystem too early and
will fail.

Add option "rootwait" in these cases. This will delay the mount
process until all such asynchronous detection tasks are done.