Join Vybrid asm-offsets.c with arch/arm/lib/asm-offsets.c
The new build system of u-boot-2014.04 has a different way of
creating the file include/generated/asm-offsets.h. There can only be
one file in arch/<arch>/lib/asm-offsets.c that is compiled and that
generates this header file. So we have to move the Vybrid declerations
from arch/arm/cpu/armv7/vybrid/asm-offsets.c to arch/arm/lib and
combine it with the existing file there.