Add fat_exists() function in F&S FAT implementation
The test command was extended to also provide -e as test for existence
of a file. This requires that a file_exists() function must be available
for all filesystems. We use our own FAT implementation and therefore
must add this function now, too. Of course we can have again wildcards
in the test.