Make U-Boot binary layout work with F&S NBoot again
The new binary layout that was introduced in u-boot-2014.07 by
moving the exception vector table from arch/arm/cpu/armv7/start.S
to arch/arm/lib/vectors.S was not recognized by NBoot anymore.
NBoot assumes a magic number 0x12345678 at offset 0x3C and jumps
to offset 0x5C where it expects the reset routine. This failed
with the new layout.

This commit modifies the layout so that this works again. It
removes the _reset variable with the pointer to the reset function
and uses a branch command on offset 0 instead. This is necessary
to have room for the magic number 0x12345678. And it introduces a
pseudo reset block at offset 0x50-0x5F that jumps to the real reset
routine so that the call from NBoot works again.