Add IOMUX_PADS() and SETUP_IOMUX_PADS() for i.MX6 Solo-X
To be able to stay as similar as possible to the code for the regular
i.MX6, we want to have macros IOMUX_PADS(), SETUP_IOMUX_PAD() and
SETUP_IOMUX_PADS(), even if there are no different settings required
at the moment (like on i.MX6 Quad/Dual and DualLite/Solo). This may
prove helpful in the future, too.