Trigger RESETOUTn at start on fsimx6sx
Some hardware must be reset manually at start, e.g. the WLAN/BT chip,
PCIe, USB hub. This is done by signal RESETOUTn that is available
on-board and also on the efusA9X SKIT. Because this is a generic reset
signal and not specific to some driver, it is most useful to do this
in U-Boot in board_init(). We use a 100ms low pulse.