Add support for config jumpers on fsimx6ul
Starting with VN27, NBoot now passes information about Ethernet,
WLAN and eMMC configuration to U-Boot. Evaluate this information
and initialize devices accordingly. This requires some rework of
the ETH and SDHC init code.

This means on efusA7UL it is now possible to have ENET1 + ENET2, or
only ENET1, or only ENET2. In the latter case, the MDIO is reconfigured
to work from ENET2, and ENET1 can remain completely off. In all other
cases MDIO is done from ENET1.

Usually SD_B (the normal sized SD slot on the efus SKIT) is started
first, and then SD_A (the Micro-SD slot on the SKIT) as second. If
WLAN is mounted, SD_A is not available. If eMMC is mounted, SD_B is
not available, but eMMC can be used as a replacement. But then eMMC is
started later than SD_A, because usually updates are not possible
from eMMC and therefore the first SD device, which is automatically
used to look for updates, should be a real slot.