In fsl_esdhc.c, invalidate dcache twice to be sure
The cache must be invalidated before the DMA for read starts. Otherwise
if the cache mode is write-back and the data cache in this region is
dirty, cache lines may be written back to RAM after the DMA has
already stored its data there and before the cache is invalidated
after the DMA has finished. In this case some data from DMA would be
lost. Just to be sure, invalidate cache twice, once before and once
after the DMA cycle.