fsl_nfc_fus.c: Return raw data in case of non-correctable ECC
On fsvybrid, in case of a non-correctably ECC error when reading a
NAND page, do not return unmodified buffers. Some callers like UBI
assume that at least the uncorrected data is loaded in this case.
Therefore do the favour and return the raw data. This makes the code
even more straightforward because we can drop a slightly complicated
optimization.

This all happens in function fus_nfc_read_page().