Reduce line length for progress in tftp and nfs
Adding the size of the already downloaded part at the end of each row
of hash marks when downloading with tftp or nfs caused the lines to be
longer than 80 characters, causing them to split on regular terminals.
This is not wanted.

To keep 64 hash marks per row, reduce the indendation on the left side
of each row from 9 to 2 blanks. This means we can not start behind
the "Loading:" string, we must start the hash marks on the next line
instead.