arch/arm/lib/vectors.S must generate relocation info
If the section ".vectors" that is declared in arch/arm/lib/vectors.S
does not have the "a" flag, then it does not create relocation
information. This means that the exception vectors will still point
to addresses before U-Boot relocation to end of RAM and all exceptions
will fail, for example the LED blinking interrupt. Therefore we need
the "a" flag, then the exception vectors will also be relocated and
point to the correct addresses.

We also add type STT_FUNC to the _start address. This will trigger
disassemby of this region as code and not data.