Move offset for NAND status more to the back on fsvybrid
Do not use offset 0x40 behind the main area for ECC status, because we
may have more than 64 bytes OOB and then the status will overwrite one
of the OOB bytes. We have four SRAM buffers with each 0x900 bytes, and
the offset must be a multiple of 8. So use offset 0x8F8, which is the
maximum possible value in each buffer. It allows for up to 248 bytes of
usable OOB area.

To reflect all the changes done on the NAND driver in the last months,
increase version string to V2.0.