Add nand convert command for fsvybrid
We originally assumed that only NAND flashes with 64 bytes OOB
area will be used on Vybrid. This assumption was wrong and also
NAND flashes with 224 bytes OOB area were used. Then the NAND
page layout was broken, but by a lucky coincidence, this worked
astonishingly stable because U-Boot and Linux used rather similar
driver code and continued to understand the data that was written
by each of them.

However the NBoot NAND driver is different and there the layout
computation resulted in a slightly different page layout on these
flashes with larger OOB area. It placed ECC and OOB data at a
different position in the page in these cases. This was no
problem as long as U-Boot was only written and updated by NBoot.
But the U-Boot MTD partition could not be read in U-Boot and
Linux because of the different layout and resulting ECC errors.
And vice versa if the U-Boot partition was written by U-Boot or
Linux, NBoot could not read this partition anymore because of the
different page layout. In this case the board did not boot
anymore!

The problem was caused by a static NAND layout here that only
worked for 64 bytes OOB and constantly used oobavail=30 for 16
correctable ECC bits. We have already fixed this behaviour in
U-Boot and Linux, by computing the layout and oobavail
dynamically at run-time after the OOB size is known. Now U-Boot,
Linux and NBoot use the same page layout for all OOB sizes.

However there are boards in the field with 224 bytes OOB that use
the old style and need to be updated to newer U-Boot and Linux
versions that use the new style. This requires converting the
whole NAND flash content from old style to new style. We do this
by reading the data in old style, erasing the NAND flash and
writing back the data in new style with the new command "nand
convert".

This means we need to add the capability to also read data in
old style. We actually try both. First we try new style and if
this fails, we also try old style. Only if this also fails, we
check for an empty page or return an uncorrectable error.

This patch does the following modifications:

- Add subcommand "convert" to the nand command in cmd_nand.c
  nand convert on|1:  switch conversion on
  nand convert off|0: switch conversion off
  nand convert addr offset size: convert nand region
  nand convert addr partition: convert MTD partition

- Add new function nand_convert_skip_bad() to nand_util.c
  This function reads the given NAND region and writes it back.
  This is done block by block. Because old data may be on old style
  and is written back in new style, this performs the conversion.

- Add support to also read data in old style to function
  fus_nfc_read_page() in fsl_nfc_fus.c. This is done if reading in
  new style fails, but only if OOB size is not 64 (here old and new
  versions are identical and an additional read does not make sense)
  and only if conversion is switched on. Activate conversion at
  start.

- Add entry convert to struct mtd_info to hold info whether conversion
  is switched on or off.

- Define CONFIG_CMD_NAND_CONVERT in fsvybrid.h to include conversion
  code in the build.