Fix ethernet on NetDCUA5 Rev 1.21
On some platforms we use CKO1 to output the RMII clock and then
the RMII clock pin is an input. This works around a silicon bug on
Vybrid where data is latched at the wrong edge of the RMII clock if
RMII clock is used as an output.

This variant has to be used if NBoot sets FEAT2_RMIICLK_CKO1 in
chFeatures2 of the NBoot Args. However this variant never worked
correctly. This is fixed by this commit.

If FEAT2_RMIICLK_CKO1 is set:
- Use PLL5 as clock source for CKO1, not RMII clock
- Do not activate RMII clock
- Switch ENET clock mux to RMII clock
- Configure CKO1 (PTB10) as output
- Configure RMII clock pin (PTA6) as input

If FEAT2_RMIICLK_CKO1 is not set:
- Switch ENET clock mux to PLL5
- Activate RMII clock
- Configure RMII clock pin (PTA6) as output