Remove fsload variable from fsimx6, fsimx6sx and fsimx6ul
The variable fsload was meant to be set to "fatload" or "ext2load".
However since the generic load command is available, this variable is
not required anymore, load can automatically use the correct version,
depending on the filesystem that was mounted.

Remove fsload from fsimx6.c, fsimx6sx.c and fsimx6ul.c.