Reduce drive strength for ethernet clock signal
On boards that output ethernet clock on CKO1 and loop it back to the
RMIICLK pin as input, the drive strength of the clock signal can be
reduced. This helps conforming to EMC tests.

On boards that use RMIICLK as output, the drive strength must stay
high to avoid ethernet problems.