Add dynamic device tree boot modifications for F&S boards
On F&S boards that use devic trees (fsimx6/fsimx6sx/fsimx6ul), some
board settings can only be detected at run-time. However they should
also be set correctly in the device tree so that Linux can start with
the right configuration.

For example the ECC strength for NAND flash on F&S boards is detected
in NBoot and passed on in the NBootArgs to U-Boot. So in the past,
U-Boot could handle ECC strength dynamically, but Linux looked at
the static device tree and probably had the wrong setting. We actually
needed different device trees, one for each possible ECC strength
value.

Now U-Boot can modify such settings in the device tree on-the-fly before
Linux is started. When CONFIG_OF_BOARD_SETUP is defined, the function
ft_board_setup() is called after the device tree is loaded to RAM just
before the Linux kernel is started. Here we can do some final fixes of
the device tree settings.

Currently the following dynamic settings are done:

- Set fus,ecc_strength in the gpmi-nand node
- Set ecc_strength, nand_state, board_name, board_revision, platform,
  arch, features1 and features2 in the bdinfo node
- Set the MAC addresses in the bdinfo node

If a value is already given in the device tree, it is kept and is not
overwritten by U-Boot. This helps keeping compatibility with existing
device trees. Or from another point of view, this allows overriding
the dynamic mechanism by giving fix values in the device tree. If this
happens, a message is shown during the boot process:

  "Keeping property ... from device tree!"

In addition to the settings above, the following is done without regard
to the existing values in the device tree:

- Deactivate all ethernet nodes that features2 reports as non-existent
  (by setting the status to "disabled")