[fsimx6/ul/sx] Do not update existing properties in the device tree
When setting the run-time properties in the device tree on fsimx6,
fsimx6ul and fsimx6sx, do not overwrite properties that are already
set in the device tree. Values of the device tree should take precedence
over the values found in U-Boot to have a possibility to override U-Boot.
The messages that were shown already told that these properties were kept,
but the values were updated nonetheless. Change this behavior and actually
keep the values in this case.