[fsimx6/sx/ul] Add reset_cause, nboot_version, u-boot_version to bdinfo
For fsimx6, fsimx6sx and fsimx6ul, add the following entries to the
bdinfo node of the device tree:

reset_cause:
The reason for the last reset as a string; this requires some
modifications in arch/arm/imx-common/cpu.c and an additional entry in
the ARM specific part of struct arch_global_data, because the reset
cause is read very early in the boot process where no regular variables
should be used.

nboot_version:
The NBoot version (four character string) as provided by NBoot in the
boot args.

u-boot_version:
The version_string[] that is also shown as first line in the version
command.