[fsimx6/sx/ul] Add support for 2 GB RAM
In the past, the device tree was relocated to the end of RAM before Linux
was booted. This caused the Linux kernel to hang if the board was equipped
with 2 GB RAM.

By setting variable fdt_high to 0xffffffff, this relocation is not done.
Our default location does not need this relocation.

By setting CONFIG_SYS_BOOTMAPSZ, we can also avoid that the device tree
is copied behind offset 0x6f800000 in RAM in case that fdt_high is not
set. This is the limit that can be directly accessed as lowmem in the
regular page table layout of Linux with a 2GB/2GB user/kernel split.

This modification is done in fsimx6.h, fsimx6sx.h and fsimx6ul.h.

While we are at it, also improve some other minor things in these files:

- Update copyright message to 2017
- Update list of supported boards
- Include FDT in flash layout comment
- Change some comments
- Remove some settings that were commented anyway