Add recovergpio to all F&S boards
By defining environment variable recovergpio, the user can define a
GPIO pin and a state (high or low). When this GPIO has the given state
at boot time (after the boot delay has counted down), the recovery check
is started instead of the update check.

The following examples show the syntax of the value: the GPIO number,
followed by an underscore, followed by "high" or "low". The GPIO number
can be taken from the GPIO Reference Card.

  setenv recovergpio 123_low
  setenv recovergpio 65_high
  setenv recovergpio 33

If just the gpio number is given like in the last example, the active
state defaults to "low".